+++ sbncode/BeamSpillInfoRetriever/CMakeLists.txt
@@ -1,14 +1,19 @@
-add_subdirectory(BNBRetriever)
+find_package(ifbeam)
+find_package(ifdh_art)
+
+include_directories(
+    ${IFBEAM_FQ_DIR}/include
+    ${IFDH_ART_INC}
+)
+
 add_subdirectory(SBNDBNBRetriever)
 add_subdirectory(SBNDBNBZEROBIASRetriever)
 add_subdirectory(SBNDBNBEXTRetriever)
 add_subdirectory(NuMIRetriever)
-add_subdirectory(BNBEXTRetriever)
 add_subdirectory(NuMIEXTRetriever)
+add_subdirectory(ICARUSBNBEXTRetriever)
+add_subdirectory(ICARUSBNBRetriever)
 add_subdirectory(job)
-
-find_package(ifbeam)
-find_package(ifdh_art)
 
 art_make_library(LIBRARIES Boost::system
         LIBRARY_NAME sbn_MWRData
@@ -28,8 +33,8 @@ art_make_library(
         sbn_MWRData
         larcorealg::CoreUtils
 	
-	LIBRARY_NAME sbn_SBNDPOTTools
-        SOURCE SBNDPOTTools.cpp
+	LIBRARY_NAME sbn_POTTools
+        SOURCE POTTools.cpp
 )
 
 art_make_library(
