+++ sbncode/BeamSpillInfoRetriever/POTTools.cpp
@@ -1,12 +1,10 @@
-#include "SBNDPOTTools.h"
+#include "POTTools.h"
 
 using namespace std;
 
-namespace sbn{
+namespace sbn::pot{
 
-  // Function for extracting the PTB information (current timestamp,
-  // previous timestamp, gate counter). Used in SBND modules only.
-  sbn::PTBInfo_t extractPTBInfo(art::Handle<std::vector<artdaq::Fragment> > cont_frags, int HLT) {
+  sbn::pot::PTBInfo_t extractPTBInfo(art::Handle<std::vector<artdaq::Fragment> > cont_frags, int HLT) {
     bool foundHLT = false;
     PTBInfo_t PTBInfo;
     for (auto const& cont : *cont_frags)
@@ -22,11 +20,11 @@ namespace sbn{
             if (ctb_frag.Trigger(word_i)->IsHLT() && ctb_frag.Trigger(word_i)->IsTrigger(HLT))
             {
               foundHLT = true;
-              uint64_t RawprevPTBTimeStamp = ctb_frag.PTBWord(word_i)->prevTS * 20;
-              uint64_t RawcurrPTBTimeStamp = ctb_frag.Trigger(word_i)->timestamp * 20;
-              double currTS_candidate = std::bitset<64>(RawcurrPTBTimeStamp/20).to_ullong()/50e6;
+              uint64_t RawprevPTBTimeStamp = ctb_frag.PTBWord(word_i)->prevTS;
+              uint64_t RawcurrPTBTimeStamp = ctb_frag.Trigger(word_i)->timestamp;
+              std::uint64_t currTS_candidate = std::bitset<64>(RawcurrPTBTimeStamp).to_ullong() * 20;
               if(currTS_candidate < PTBInfo.currPTBTimeStamp){
-                PTBInfo.prevPTBTimeStamp = std::bitset<64>(RawprevPTBTimeStamp / 20).to_ullong()/50e6;
+                PTBInfo.prevPTBTimeStamp = std::bitset<64>(RawprevPTBTimeStamp).to_ullong() * 20;
                 PTBInfo.currPTBTimeStamp = currTS_candidate;
                 PTBInfo.GateCounter = ctb_frag.Trigger(word_i)->gate_counter;
               }
@@ -45,7 +43,35 @@ namespace sbn{
     }
   }
 
-  // Function for extracting the TDC timestamp. Used in SBND modules only.
+  std::vector<PTBInfo_t> extractAllPTBInfo(std::vector<artdaq::Fragment> const& cont_frags) {
+    std::vector<PTBInfo_t> PTBInfoVec;
+    for (auto const& cont : cont_frags)
+    {
+      artdaq::ContainerFragment cont_frag(cont);
+      for (size_t fragi = 0; fragi < cont_frag.block_count(); ++fragi)
+      {
+        artdaq::Fragment frag = *cont_frag[fragi];
+        sbndaq::CTBFragment ctb_frag(frag);   
+        for(size_t word_i = 0; word_i < ctb_frag.NWords(); ++word_i)
+        {
+          if(ctb_frag.Trigger(word_i)){
+            PTBInfo_t PTBInfo;
+            uint64_t RawprevPTBTimeStamp = ctb_frag.PTBWord(word_i)->prevTS;
+            uint64_t RawcurrPTBTimeStamp = ctb_frag.Trigger(word_i)->timestamp;
+            PTBInfo.prevPTBTimeStamp = std::bitset<64>(RawprevPTBTimeStamp).to_ullong() * 20;
+            PTBInfo.currPTBTimeStamp = std::bitset<64>(RawcurrPTBTimeStamp).to_ullong() * 20;
+            PTBInfo.GateCounter = ctb_frag.Trigger(word_i)->gate_counter;
+            PTBInfo.isHLT = ctb_frag.Trigger(word_i)->IsHLT();
+            PTBInfo.triggerWord = ctb_frag.Trigger(word_i)->trigger_word;
+            PTBInfoVec.push_back(PTBInfo);
+          }
+        } 
+      } 
+    } 
+  
+    return PTBInfoVec;
+  }
+
   double extractTDCTimeStamp(art::Handle<std::vector<artdaq::Fragment> > cont_frags) {
   
     double TDCTimeStamp = 0;
@@ -62,8 +88,6 @@ namespace sbn{
     return TDCTimeStamp;
   }
 
-  // Collect device information to make BNBSpillInfo object. Used in
-  // ICARUS and SBND modules.
   sbn::BNBSpillInfo makeBNBSpillInfo
   (art::EventID const& eventID, double time, MWRdata_t const& MWRdata, std::vector<int> const& matched_MWR, std::unique_ptr<ifbeam_ns::BeamFolder> const& bfp,  const std::unique_ptr<ifbeam_ns::BeamFolder> & offsets,const  std::unique_ptr<ifbeam_ns::BeamFolder> & vp873)
   {
@@ -78,6 +102,8 @@ namespace sbn{
     double LM875A = -999; // units R/s
     double LM875B = -999; // units R/s
     double LM875C = -999; // units R/s
+    double HP873 = -999; // units mm
+    double VP873 = -999; // units mm; not in the first IFBeam query bunch
     double HP875 = -999; // units mm
     double VP875 = -999; // units mm
     double HPTG1 = -999; // units mm
@@ -86,55 +112,66 @@ namespace sbn{
     double VPTG2 = -999; // units mm
     double BTJT2 = -999; // units Deg C
     double THCURR = -999; // units kiloAmps
+    double M875HS = -999; // units mm
+    double M875VS = -999; // units mm
+    double M875HM = -999; // units mm
+    double M875VM = -999; // units mm
+    double M876HS = -999; // units mm
+    double M876VS = -999; // units mm
+    double M876HM = -999; // units mm
+    double M876VM = -999; // units mm
     
 
-    double VP873 = -999; //units mm; not in the first IFBeam query bunch
-    double HP875Offset = -999;//units mm; make a another separate IFBeam query bunch for offsets
-    double VP875Offset = -999;//units mm
-    double VP873Offset = -999;//units mm
-    double HPTG1Offset = -999;//units mm
-    double HPTG2Offset = -999;//units mm
-    double VPTG1Offset = -999;//units mm
-    double VPTG2Offset = -999;//units mm
+    double HP875Offset =-999;//units mm; make a another separate IFBeam query bunch for offsets
+    double VP875Offset =-999;//units mm
+    double VP873Offset =-999;//units mm
+    double HPTG1Offset =-999;//units mm
+    double HPTG2Offset =-999;//units mm
+    double VPTG1Offset =-999;//units mm
+    double VPTG2Offset =-999;//units mm
 
     double TOR860_time = -999; // units s
 
-    double FOM = -999; 
+    double FOM =-999; 
 
     // Here we request all the devices
     // since sometimes devices fail to report we'll
     // allow each to throw an exception but still move forward
     // interpreting these failures will be part of the beam quality analyses
-  
-    try{bfp->GetNamedData(time, "E:TOR860@",&TOR860,&TOR860_time);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:TOR875",&TOR875);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:LM875A",&LM875A);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:LM875B",&LM875B);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:LM875C",&LM875C);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:HP875",&HP875);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:VP875",&VP875);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:HPTG1",&HPTG1);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:VPTG1",&VPTG1);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:HPTG2",&HPTG2);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:VPTG2",&VPTG2);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:BTJT2",&BTJT2);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{bfp->GetNamedData(time, "E:THCURR",&THCURR);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
 
+    try{bfp->GetNamedData(time, "E:TOR860@",&TOR860,&TOR860_time);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:TOR875",&TOR875);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:LM875A",&LM875A);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:LM875B",&LM875B);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:LM875C",&LM875C);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:HP873",&HP873);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{vp873->GetNamedData(time, "E:VP873",&VP873);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:HP875",&HP875);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:VP875",&VP875);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:HPTG1",&HPTG1);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:VPTG1",&VPTG1);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:HPTG2",&HPTG2);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:VPTG2",&VPTG2);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:BTJT2",&BTJT2);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:THCURR",&THCURR);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
 
-    
-    try{vp873->GetNamedData(time, "E:VP873",&VP873);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    
-    try{offsets->GetNamedData(time, "E_VP873S",&VP873Offset);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{offsets->GetNamedData(time, "E_HP875S",&HP875Offset);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{offsets->GetNamedData(time, "E_VP875S",&VP875Offset);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{offsets->GetNamedData(time, "E_HPTG1S",&HPTG1Offset);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{offsets->GetNamedData(time, "E_HPTG2S",&HPTG2Offset);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{offsets->GetNamedData(time, "E_VPTG1S",&VPTG1Offset);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
-    try{offsets->GetNamedData(time, "E_VPTG2S",&VPTG2Offset);}catch (WebAPIException &we) {mf::LogDebug("SBNDBNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:M875HS",&M875HS);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:M875VS",&M875VS);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:M875HM",&M875HM);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:M875VM",&M875VM);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:M876HS",&M876HS);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:M876VS",&M876VS);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:M876HM",&M876HM);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{bfp->GetNamedData(time, "E:M876VM",&M876VM);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";} 
 
+    try{offsets->GetNamedData(time, "E_VP873S",&VP873Offset);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{offsets->GetNamedData(time, "E_HP875S",&HP875Offset);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{offsets->GetNamedData(time, "E_VP875S",&VP875Offset);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{offsets->GetNamedData(time, "E_HPTG1S",&HPTG1Offset);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{offsets->GetNamedData(time, "E_HPTG2S",&HPTG2Offset);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{offsets->GetNamedData(time, "E_VPTG1S",&VPTG1Offset);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
+    try{offsets->GetNamedData(time, "E_VPTG2S",&VPTG2Offset);}catch (WebAPIException &we) {mf::LogDebug("BNBRetriever")<< "At time : " << time << " " << "got exception: " << we.what() << "\n";}
 
-
-  
     //crunch the times 
     unsigned long int time_closest_int = (int) TOR860_time;
     double time_closest_ns = (TOR860_time - time_closest_int)*1e9;
@@ -146,6 +183,8 @@ namespace sbn{
     beamInfo.LM875A = LM875A;
     beamInfo.LM875B = LM875B;
     beamInfo.LM875C = LM875C;
+    beamInfo.HP873 = HP873;
+    beamInfo.VP873 = VP873;
     beamInfo.HP875 = HP875;
     beamInfo.VP875 = VP875;
     beamInfo.HPTG1 = HPTG1;
@@ -154,6 +193,14 @@ namespace sbn{
     beamInfo.VPTG2 = VPTG2;
     beamInfo.BTJT2 = BTJT2;
     beamInfo.THCURR = THCURR;
+    beamInfo.M875HS = M875HS;
+    beamInfo.M875VS = M875VS;
+    beamInfo.M875HM = M875HM;
+    beamInfo.M875VM = M875VM;
+    beamInfo.M876HS = M876HS;
+    beamInfo.M876VS = M876VS;
+    beamInfo.M876HM = M876HM;
+    beamInfo.M876VM = M876VM;
     beamInfo.spill_time_s = time_closest_int;
     beamInfo.spill_time_ns = time_closest_ns;    
     beamInfo.VP873 = VP873;
@@ -166,8 +213,7 @@ namespace sbn{
     beamInfo.VPTG2Offset = VPTG2Offset;
     beamInfo.FOM = FOM;
 
-       
-  
+    
     for(auto const& MWRdata: unpacked_MWR){
       std::ignore = MWRdata;
       assert(!MWRdata.empty());
@@ -212,9 +258,6 @@ namespace sbn{
     return beamInfo;
   }
 
-  // Returns true if a time in IFBeam corresponds to a broken clock false if
-  // it is a real spill. Check this by seeing if all devices fire at once or
-  // if time is just a single outlier.
   bool BrokenClock(double time, std::unique_ptr<ifbeam_ns::BeamFolder> const& bfp)
   {
     double TOR860 = -999; // units e12 protons
@@ -263,8 +306,7 @@ namespace sbn{
     }
   }
 
-  // extract spill times from the IFBeam data base. Used in SBND and ICARUS.
-  sbn::MWRdata_t extractSpillTimes(TriggerInfo_t const& triggerInfo, std::unique_ptr<ifbeam_ns::BeamFolder> const& bfp, std::unique_ptr<ifbeam_ns::BeamFolder> const& bfp_mwr, double fTimePad, double MWRtoroidDelay, sbn::MWRData mwrdata) {
+  sbn::pot::MWRdata_t extractSpillTimes(TriggerInfo_t const& triggerInfo, std::unique_ptr<ifbeam_ns::BeamFolder> const& bfp, std::unique_ptr<ifbeam_ns::BeamFolder> const& bfp_mwr, double fTimePad, double MWRtoroidDelay, sbn::MWRData mwrdata) {
     
     // These lines get everything primed within the IFBeamDB.
     try{bfp->FillCache((triggerInfo.t_current_event)+fTimePad);} catch (WebAPIException &we) {};     
@@ -375,6 +417,3 @@ namespace sbn{
   }
 
 }
-
-
-
