+++ sbncode/BeamSpillInfoRetriever/SBNDBNBRetriever/CMakeLists.txt
@@ -3,7 +3,7 @@ find_package(ifdh_art)
 
 cet_build_plugin(SBNDBNBRetriever art::module
     LIBRARIES
-        sbn_SBNDPOTTools
+        sbn_POTTools
 	sbn_getFOM
 )
